test(compactor): `test_lsm_major_compactor` moved to Compactor and added Leveled structures after checking Major Compaction
